% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_genes.R
\name{filter_low_genes}
\alias{filter_low_genes}
\title{Filter out genes with low expression in RNAseq data}
\usage{
filter_low_genes(df, min_expr = 10, percent_cutoff = 90, metric = "units")
}
\arguments{
\item{df}{expression data. A data frame in which columns are genes and
rows are samples.}

\item{min_expr}{minimum expression. The minimum value that a gene must be
expressed across the percentage of samples selected. If no user input is
provided the default is 10.}

\item{percent_cutoff}{percentage cutoff. The percent of samples that minimum
gene expression (min_expr) is required. If no user input is provided the
default is 90.}

\item{metric}{metric/type of data. The type of data (e.g. FPKM, RSEM,
raw count, etc.) the expression data is formatted in. This does not change
how analysis is done, it simply makes the summary message of the function
relevant to your specific analysis. If no user input is provided the the
default is "unit".}
}
\description{
This function checks data for individual genes that express below a certain
threshold across a given percentage of samples, filters them out, and returns
a data frame with the genes that successfully met the user-provided criteria.
Upon completion, a message summarizing how many genes were removed is also
output.
}
